+++ src/google-sheets/service/google-sheets.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { GoogleSheetsRepository } from "../domain/repository/google-sheets.repository";
-import { AddValue } from "src/lib/add-value.type";
+import { DateTime } from "src/lib/types/datetime/datetime.type";
+import { SHEETS_NAMES } from "src/constants/sheets-name/sheets-name";
 
 
 @Injectable()
@@ -9,7 +10,7 @@ export class GoogleSheetsService {
     private readonly googleSheetsRepository: GoogleSheetsRepository
   ){}
 
-  async appendRow(range: string, values: AddValue) {
+  async appendRow(range: string, values: DateTime) {
     try {
       await this.googleSheetsRepository.appendRow(range, values);
     } catch (error) {
@@ -17,9 +18,9 @@ export class GoogleSheetsService {
     }
   }
 
-  async getLasRowValue(range: string): Promise<string> {
+  async getLastRowValue(range: string): Promise<string> {
     try {
-      const data = await this.googleSheetsRepository.getLasRowValue(range);
+      const data = await this.googleSheetsRepository.getLastRowValue(range);
       return data;
     } catch (error) {
       this.googleSheetsRepository.failure(error);
@@ -32,6 +33,17 @@ export class GoogleSheetsService {
 
       const dateExists = data.some(row => row[0] === date);
       return dateExists;
+    } catch (error) {
+      this.googleSheetsRepository.failure(error);
+    }
+  }
+
+  async getDate(date:string, time:string){
+    try {
+      const data = await this.googleSheetsRepository.getDates(`${SHEETS_NAMES[0]}!A:C`);
+      
+      const index = data.findIndex(row => row[0] === date && row[1] === time && (!row[2] || row[2].trim() === "")) + 1;
+      return index;
     } catch (error) {
       this.googleSheetsRepository.failure(error);
     }